fix interface 제거
