+++ src/main/java/com/flab/goodchoice/item/domain/ItemEntryService.java
@@ -1,7 +1,22 @@
 package com.flab.goodchoice.item.domain;
 
-public interface ItemEntryService {
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-    String registerItem(ItemCommand.RegisterItemRequest request);
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class ItemEntryService {
+
+    private final ItemStore itemStore;
+
+    @Transactional
+    public String registerItem(ItemCommand.RegisterItemRequest command) {
+        var initItem = command.toEntity();
+        var item = itemStore.store(initItem);
+        return item.getItemToken();
+    }
 
 }
