+++ src/main/java/com/flab/goodchoice/item/domain/ItemListService.java
@@ -1,10 +1,19 @@
 package com.flab.goodchoice.item.domain;
 
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
 import java.util.List;
 
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class ItemListService {
 
-public interface ItemListService {
+    private final ItemReader itemReader;
 
-    List<Item> retrieveAllItemInfo();
-
+    public List<Item> retrieveAllItemInfo() {
+        return itemReader.getItems();
+    }
 }
