--- src/main/java/com/flab/goodchoice/item/domain/ItemModifyServiceImpl.java
@@ -1,30 +1,0 @@
-package com.flab.goodchoice.item.domain;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class ItemModifyServiceImpl implements ItemModifyService {
-
-    private final ItemReader itemReader;
-
-    @Override
-    @Transactional
-    public void changeEndOfSale(String itemToken) {
-        var item = itemReader.getItemBy(itemToken);
-        item.changeEndOfSale();
-    }
-
-    @Override
-    @Transactional
-    public ItemInfo.Main updateItem(String itemToken, ItemCommand.UpdateItemRequest request) {
-        var item = itemReader.getItemBy(itemToken);
-        item.update(request.getItemName(), request.getItemPrice(), request.getItemStock());
-        return new ItemInfo.Main(item);
-    }
-
-}
